@@ -2,66 +2,66 @@
   "name": "4x-archive",
   "version": "0.0.1",
   "scripts": {
-    "dev": "NODE_PATH=./src dotenv -- nodemon -e ts,graphql -x ts-node src/index.ts",
-    "debug": "NODE_PATH=./src dotenv -- nodemon -e ts,graphql -x ts-node --inspect src/index.ts",
+    "dev": "dotenv -- nodemon src/index.ts",
     "build": "rimraf dist && tsc",
-    "doc": "graphdoc -e http://localhost:4000/ -o ./doc/schema",
-    "lint": "eslint -c .eslintrc.js --ext .ts src",
-    "knex": "NODE_PATH=./src ts-node ./node_modules/.bin/knex"
+    "lint": "eslint --ext .js,.ts -c .eslintrc.js src/",
+    "prettier": "prettier --config .prettierrc.js --write --check .",
+    "knex": "dotenv -- knex",
+    "test": "dotenv -- ava"
   },
   "dependencies": {
-    "@2fd/graphdoc": "^2.4.0",
-    "apollo-server-express": "^2.15.0",
-    "async-mutex": "^0.2.3",
+    "apollo-server-express": "^3.7.0",
+    "async-mutex": "^0.3.2",
     "bcryptjs": "2.4.3",
-    "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
-    "dataloader": "^2.0.0",
-    "express": "^4.17.1",
-    "express-graphql": "^0.9.0",
-    "file-type": "^14.6.2",
+    "cookie-parser": "^1.4.6",
+    "dataloader": "^2.1.0",
+    "express": "^4.18.1",
+    "file-type": "^17.1.1",
     "fluent-ffmpeg": "^2.1.2",
-    "fs-capacitor": "^6.2.0",
-    "fs-jetpack": "^2.4.0",
-    "graphql": "14.5.8",
-    "graphql-playground-middleware-express": "^1.7.18",
-    "graphql-postgres-subscriptions": "^1.0.5",
-    "graphql-relay": "^0.6.0",
-    "graphql-subscriptions": "^1.1.0",
-    "graphql-upload": "^11.0.0",
-    "hashids": "^2.2.1",
-    "knex": "^0.21.1",
+    "fs-capacitor": "^7.0.1",
+    "fs-jetpack": "^4.3.1",
+    "graphql": "16.4.0",
+    "graphql-playground-middleware-express": "^1.7.23",
+    "graphql-relay": "^0.10.0",
+    "graphql-subscriptions": "^2.0.0",
+    "graphql-upload": "^13.0.0",
+    "graphql-ws": "^5.8.1",
+    "hashids": "^2.2.10",
+    "knex": "^2.0.0",
     "morgan": "^1.10.0",
-    "node-vibrant": "^3.1.5",
-    "objection": "^2.2.0",
-    "objection-unique": "^1.2.1",
-    "pg": "^8.2.1",
+    "node-vibrant": "^3.1.6",
+    "objection": "^3.0.1",
+    "objection-unique": "^1.2.2",
+    "pg": "^8.7.3",
     "pg-escape": "^0.2.0",
-    "pg-ipc": "^1.0.4",
-    "sharp": "^0.25.4",
-    "sodium": "^3.0.2",
-    "string-strip-html": "^4.5.0",
-    "subscriptions-transport-ws": "^0.9.16",
-    "telegraf": "^3.38.0",
+    "pg-ipc": "^1.0.5",
+    "sharp": "^0.30.4",
+    "string-strip-html": "^8.3.0",
+    "telegraf": "^4.8.1",
     "tmp": "^0.2.1"
   },
   "devDependencies": {
     "@types/bcryptjs": "2.4.2",
-    "@types/body-parser": "^1.19.0",
-    "@types/express": "4.17.3",
-    "@types/graphql-upload": "^8.0.3",
-    "@types/knex": "^0.16.1",
-    "@types/morgan": "^1.9.1",
-    "@types/node": "14.0.13",
-    "@types/sharp": "^0.25.0",
-    "@types/tmp": "^0.2.0",
-    "@typescript-eslint/eslint-plugin": "^3.3.0",
-    "@typescript-eslint/parser": "^3.3.0",
-    "dotenv-cli": "3.1.0",
-    "eslint": "^7.3.0",
-    "eslint-plugin-jsdoc": "^27.1.2",
-    "nodemon": "2.0.4",
-    "ts-node": "8.10.2",
-    "typescript": "3.9.5"
+    "@types/body-parser": "^1.19.2",
+    "@types/cookie-parser": "^1.4.3",
+    "@types/express": "4.17.13",
+    "@types/graphql-upload": "^8.0.11",
+    "@types/morgan": "^1.9.3",
+    "@types/node": "17.0.31",
+    "@types/sharp": "^0.30.2",
+    "@types/tmp": "^0.2.3",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
+    "dotenv-cli": "5.1.0",
+    "eslint": "^8.15.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-jsdoc": "^39.2.9",
+    "nodemon": "2.0.16",
+    "prettier": "^2.6.2",
+    "prettier-plugin-jsdoc": "^0.3.38",
+    "ts-node": "10.7.0",
+    "tsconfig-paths": "^3.12.0",
+    "typescript": "4.6.4"
   }
 }
