@@ -1,12 +1,11 @@
+exports.up = async (knex) => {
+  await knex.schema.alterTable('User', (table) => {
+    table.boolean('darkmode').defaultTo(false).notNullable()
+  })
+}
 
-exports.up = async knex => {
-    await knex.schema.alterTable('User', table => {
-        table.boolean('darkmode').defaultTo(false).notNullable()
-    })
-};
-
-exports.down = async knex => {
-    await knex.schema.alterTable('User', table => {
-        table.dropColumn('darkmode')
-    })
-};
+exports.down = async (knex) => {
+  await knex.schema.alterTable('User', (table) => {
+    table.dropColumn('darkmode')
+  })
+}
