@@ -1,12 +1,11 @@
+exports.up = async (knex) => {
+  await knex.schema
+    .raw("ALTER TYPE format ADD VALUE 'AUDIO';")
+    .raw("ALTER TYPE format ADD VALUE 'TEXT';")
+}
 
-exports.up = async knex => {
-    await knex.schema
-        .raw('ALTER TYPE format ADD VALUE \'AUDIO\';')
-        .raw('ALTER TYPE format ADD VALUE \'TEXT\';');
-};
+exports.down = async (knex) => {
+  // no.
+}
 
-exports.down = async knex => {
-    // no.
-};
-
-exports.config = { transaction: false };
+exports.config = { transaction: false }
