@@ -1,17 +1,18 @@
 import { GraphQLFieldConfig, GraphQLList, GraphQLNonNull } from 'graphql'
-import Context from 'Context'
+import Context from '@src/Context'
 
 import SessionType from './SessionType'
-import SessionActions from 'actions/SessionActions'
+import SessionActions from '@src/actions/SessionActions'
 
 const userSessions: GraphQLFieldConfig<any, any, any> = {
-    description: 'Returns a list of sessions of the the currently authenticated user.',
-    type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(SessionType))),
-    resolve: async (parent, args, context: Context, resolveInfo) => {
-        return SessionActions.qGetUserSessions(context)
-    },
+  description:
+    'Returns a list of sessions of the the currently authenticated user.',
+  type: new GraphQLNonNull(new GraphQLList(new GraphQLNonNull(SessionType))),
+  resolve: async (parent, args, context: Context, resolveInfo) => {
+    return SessionActions.qGetUserSessions(context)
+  },
 }
 
 export default {
-    userSessions,
+  userSessions,
 }
