@@ -1,7 +1,7 @@
-import RequestError from './RequestError'
+import { ApolloError } from 'apollo-server-core'
 
-export default class AuthorizationError extends RequestError {
-    constructor(msg?: string) {
-        super('You are not authorized to do this.', msg)
-    }
+export default class AuthorizationError extends ApolloError {
+  constructor(msg?: string) {
+    super(msg || 'You are not authorized to do this.', 'AUTHORIZATION_ERROR')
+  }
 }
