@@ -1,7 +1,7 @@
-import RequestError from './RequestError'
+import { ApolloError } from 'apollo-server-core'
 
-export default class NotFoundError extends RequestError {
-    constructor(msg?: string) {
-        super('Item not found.', msg)
-    }
+export default class NotFoundError extends ApolloError {
+  constructor(msg?: string) {
+    super(msg || 'Item not found.', 'NOT_FOUND_ERROR')
+  }
 }
