@@ -1,7 +1,6 @@
 export { default as AuthenticationError } from './AuthenticationError'
 export { default as AuthorizationError } from './AuthorizationError'
-export { default as FileStorageError } from './FileStorageError'
 export { default as InputError } from './InputError'
+export { default as ValidationInputError } from './ValidationInputError'
 export { default as NotFoundError } from './NotFoundError'
 export { default as RequestError } from './RequestError'
-export { default as UploadError } from './UploadError'
